XenGnttabDmaBufferMapper add new class to handle map ioctls
Add new class which should work with the following calls:
xengnttab_dmabuf_map_refs_to_buf
xengnttab_dmabuf_map_release
xengnttab_dmabuf_map_wait_released

Those calls are intended to map existing dma_buf, provided with
the descriptor to grant refs. It will be useful when there is no
support for the default backing storags, which is created by
GntdevExporter, so external dmabuf should be provided to be
used as the backing storage.

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>